requirements.txt
discord
aiohttp
asyncio
Pillow